+++ src/main/java/com/formation/project/computerDatabase/dao/IStatsDao.java
@@ -1,6 +1,6 @@
 package com.formation.project.computerDatabase.dao;
 
 public interface IStatsDao {
-	void logOperation(Integer computerId, String statType);
+	void logOperation(Long computerId, String statType);
 	
 }
